Split `test_serializers.py` into package grouped by use case
